Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float d;
-    scanf("%f",&d);
-    printf("You entered: %f",d);
+    double d;
+    scanf("%lf",&d);
+    printf("You entered: %.4lf",d);
     return 0;
 }
